# store/openapi.yaml
openapi: 3.1.0
info:
  title: Brain store contract
  version: "1.1"
  summary: The HTTP interface between a Brain client and a remote store.
  description: >
    Machine-readable companion to CONTRACT.md. Where the two disagree,
    CONTRACT.md is correct.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://store.example
    description: Your store
security:
  - bearer: []
paths:
...
  /auth/me:
    get:
      summary: Describe the caller
      responses:
        "200":
          description: The authenticated user.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  storage_used: { type: integer, description: Bytes stored for this user. }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/brains:
    get:
      summary: List the caller's brains, oldest first
      responses:
        "200":
          description: Brain records.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Brain" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Create a brain
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, default: default }
      responses:
        "201":
          description: The new brain.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Brain" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
...
  schemas:
...
    User:
      type: object
      required: [id, email]
      properties:
        id: { type: string }
        email: { type: string }
      additionalProperties: true
    Brain:
      type: object
      required: [id, name, size_bytes, file_count, checksum, last_synced_at, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        size_bytes: { type: integer }
        file_count: { type: integer, description: Regular-file entries in the archive. }
        checksum: { type: [string, "null"], description: Lowercase hex SHA-256 of the archive. }
        last_synced_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
      additionalProperties: true
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: Human-readable. Do not parse. }
      additionalProperties: true
  responses:
    Error:
      description: The request failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, unknown or expired token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...